Add files via upload
You are given a group of people, the type of the group, and the day of the week they are going to stay. Based on
that information calculate how much they have to pay and print that price on the console. Use the table below. In
each cell is the price for a single person.
The output should look like that: `Total price: {price}`.The price should be formatted to the second
decimal point.